+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/login/LoginActivity.kt
@@ -1,155 +1,24 @@
 package com.nbcfinalteam2.ddaraogae.presentation.ui.login
 
-import android.content.DialogInterface
-import android.content.Intent
 import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.activity.enableEdgeToEdge
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.activity.viewModels
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.updatePadding
-import androidx.lifecycle.flowWithLifecycle
-import androidx.lifecycle.lifecycleScope
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.common.api.ApiException
 import com.nbcfinalteam2.ddaraogae.R
-import com.nbcfinalteam2.ddaraogae.databinding.ActivityLogInBinding
-import com.nbcfinalteam2.ddaraogae.presentation.ui.main.MainActivity
-import com.nbcfinalteam2.ddaraogae.presentation.util.ToastMaker
+import com.nbcfinalteam2.ddaraogae.databinding.ActivityLoginBinding
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class LoginActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityLogInBinding
-    private val viewModel: LoginViewModel by viewModels()
-
-    private lateinit var googleSignInClient: GoogleSignInClient
-    private val activityResultLauncher =
-        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-            if (result.resultCode == RESULT_OK) {
-                val task = GoogleSignIn.getSignedInAccountFromIntent(result.data)
-                try {
-                    val account = task.getResult(ApiException::class.java)!!
-                    Log.d(TAG, "firebaseAuthWithGoogle:" + account.id)
-                    firebaseAuthWithGoogle(account.idToken!!)
-                } catch (e: ApiException) {
-                    Log.w(TAG, "구글 로그인에 실패했습니다.", e)
-                    ToastMaker.make(this, getString(R.string.login_google_sign_in_fail))
-                }
-            } else {
-                Log.e(TAG, "Google Result Error ${result}")
-                ToastMaker.make(this, getString(R.string.login_google_sign_in_fail))
-            }
-        }
-
+    private lateinit var binding:ActivityLoginBinding
     override fun onCreate(savedInstanceState: Bundle?) {
-        installSplashScreen().apply {
-            setKeepOnScreenCondition { viewModel.isLoading.value }
-        }
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        binding = ActivityLogInBinding.inflate(layoutInflater)
-        val view = binding.root
-        setContentView(view)
-        uiSetting()
-        checkIsPossible()
-        clickLoginButton()
-        viewModel.getCurrentUser()
+        binding = ActivityLoginBinding.inflate(layoutInflater)
+
+        setFragment()
+        setContentView(binding.root)
     }
-
-    private fun uiSetting() {
-        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { view, windowInsets ->
-            val insets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars())
-            view.setPadding(insets.left, insets.top, insets.right, insets.bottom)
-            windowInsets
-        }
-    }
-
-    private fun checkIsPossible() {
-        lifecycleScope.launch {
-            viewModel.userState.flowWithLifecycle(lifecycle)
-                .collect { state ->
-                    if (state == 0) successLogIn()
-                    if (state == 1) viewModel.checkVerified()
-                    if (state == 2) Toast.makeText(this@LoginActivity, R.string.login_fail, Toast.LENGTH_SHORT).show()
-                    if (state == 3) sendEmail()
-                    if (state == 98) Toast.makeText(
-                        this@LoginActivity, R.string.login_ioexception, Toast.LENGTH_SHORT).show()
-                    if (state == 99) Toast.makeText(
-                        this@LoginActivity, R.string.login_unknown_error, Toast.LENGTH_SHORT).show()
-                }
-        }
-    }
-
-    private fun sendEmail() {
-        val builder = AlertDialog.Builder(this)
-        builder.setTitle(R.string.login_dialog_title)
-            .setMessage(R.string.login_dialog_message)
-            .setPositiveButton(R.string.login_dialog_ok, DialogInterface.OnClickListener { _, _ ->
-                viewModel.sendEmail()
-                Toast.makeText(this@LoginActivity, R.string.login_send_email, Toast.LENGTH_SHORT).show()
-                viewModel.initState()
-            })
-            .setNegativeButton(R.string.login_dialog_no, DialogInterface.OnClickListener { _, _ ->
-                viewModel.deleteAccount()
-                Toast.makeText(this@LoginActivity, R.string.login_account_delete, Toast.LENGTH_SHORT).show()
-            })
-            .setCancelable(false)
-        builder.show()
-    }
-
-    private fun clickLoginButton() = with(binding) {
-        //click LoginButton
-        btLogin.setOnClickListener {
-            val email = etLoginEmail.text.toString().trim()
-            val password = etLoginPassword.text.toString().trim()
-            if(email.isBlank() || password.isBlank()) Toast.makeText(this@LoginActivity, R.string.login_input_account, Toast.LENGTH_SHORT).show()
-            else viewModel.signInEmail(email, password)
-        }
-
-        //click google Login Button
-        ibtLoginGoogle.setOnClickListener {
-            val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build()
-            googleSignInClient = GoogleSignIn.getClient(this@LoginActivity, gso)
-
-            val signInIntent = googleSignInClient.signInIntent
-            activityResultLauncher.launch(signInIntent)
-        }
-
-        //click SignUp Button
-        tvLoginSignup.setOnClickListener {
-            etLoginEmail.setText(null)
-            etLoginPassword.setText(null)
-            startActivity(Intent(this@LoginActivity, SignUpActivity::class.java))
-            finish()
-
-        }
-    }
-
-    //firebase에 값 넘겨주기, 로그인
-    private fun firebaseAuthWithGoogle(idToken: String) {
-        viewModel.signInGoogle(idToken)
-    }
-
-    private fun successLogIn() {
-        Toast.makeText(this@LoginActivity, R.string.login_success, Toast.LENGTH_SHORT).show()
-        startActivity(Intent(this, MainActivity::class.java))
-        finish()
-    }
-
-    companion object {
-        private const val TAG = "GoogleActivity"
+    private fun setFragment(){
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.activity_login, LoginFragment())
+            .commit()
     }
 }
